Module 3/birdclef2/requirements.txt
pandas
python-dotenv
numpy
lightning
scikit-learn
pydantic
soundfile
timm
wandb